RSESPRT-59: Rename word replacement class to reflect sales tracking. Add upgrader to add the option value
