+++ CRM/Prospect/Upgrader.php
@@ -7,7 +7,7 @@ use CRM_Prospect_Setup_CreateProspectWorkflowCaseStatuses as CreateProspectWorkf
 use CRM_Prospect_Setup_CreateProspectWorkflowCaseType as CreateProspectWorkflowCaseType;
 use CRM_Prospect_Setup_MoveCustomFieldsToWorkFlowCaseType as MoveCustomFieldsToWorkFlowCaseType;
 use CRM_Prospect_Setup_ProcessProspectCategoryForCustomGroupSupport as ProcessProspectCategoryForCustomGroupSupport;
-use CRM_Prospect_Setup_AddProspectCategoryWordReplacement as AddProspectCategoryWordReplacement;
+use CRM_Prospect_Setup_AddSalesOpportunityTrackingWordReplacement as AddSalesOpportunityTrackingWordReplacement;
 use CRM_Prospect_Setup_EnableRequiredComponents as EnableRequiredComponents;
 use CRM_Prospect_Uninstall_DeleteInstalledCustomGroups as DeleteInstalledCustomGroups;
 use CRM_Prospect_Uninstall_DeleteProspectMenus as DeleteProspectMenus;
@@ -50,7 +50,7 @@ class CRM_Prospect_Upgrader extends CRM_Prospect_Upgrader_Base {
     $steps = [
       new EnableRequiredComponents(),
       new CreateProspectingOptionValue(),
-      new AddProspectCategoryWordReplacement(),
+      new AddSalesOpportunityTrackingWordReplacement(),
       new CreateProspectMenus(),
       new CreateProspectWorkflowCaseStatuses(),
       new CreateProspectOwnerRelationship(),
